[fix] balance, accountName int -> long change

+++ mobile2/src/main/java/sopt/mobile2/domain/Account.java
@@ -18,9 +18,9 @@ public class Account {
 
     private String accountName;
 
-    private int balance;
+    private long balance;
 
-    private int accountNumber;
+    private long accountNumber;
 
     private String imgUrl;
