+++ mobile2/src/main/java/sopt/mobile2/dto/AccountInfoResponseDto.java
@@ -2,7 +2,7 @@ package sopt.mobile2.dto;
 
 import sopt.mobile2.domain.Account;
 
-public record AccountInfoResponseDto(String accountName, int balance, int accountNumber) {
+public record AccountInfoResponseDto(String accountName, long balance, long accountNumber) {
 
     public static AccountInfoResponseDto of(Account account) {
         return new AccountInfoResponseDto(account.getAccountName(), account.getBalance(), account.getAccountNumber());
