+++ mobile2/src/main/java/sopt/mobile2/dto/PaymentResponse.java
@@ -1,7 +1,7 @@
 package sopt.mobile2.dto;
 
 public record PaymentResponse(
-        int month, int payment
+        int month, long payment
 ) {
 }
