+++ mobile2/src/main/java/sopt/mobile2/dto/RecentTransferResponse.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 
 public record RecentTransferResponse(
         String accountName,
-        int accountNumber,
+        long accountNumber,
         boolean isAccountLike,
         LocalDateTime createdAt,
         String bankName,
